+++ Modules/Admin/Http/Controllers/CouponController.php
@@ -7,12 +7,14 @@ use Modules\Admin\Entities\Coupon;
 use App\Http\Controllers\Controller;
 use App\Http\Responses\MessageResponse;
 use App\Actions\ValidateCouponPromocode;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use Modules\Admin\Http\Requests\CouponRequest;
 use Modules\Admin\Transformers\CouponResource;
 use Modules\Admin\Http\Requests\UpdateCouponRequest;
 
 class CouponController extends Controller
 {
+    use AuthorizesRequests;
     protected $storeService, $store;
 
     public function __construct(StoreService $storeService)
@@ -23,6 +25,7 @@ class CouponController extends Controller
 
     public function index()
     {
+        $this->authorize('View-Coupon');
         $coupons = $this->store->coupons()->useFilters()->dynamicPaginate();
 
         return new MessageResponse('coupons', CouponResource::collection($coupons));
@@ -30,6 +33,7 @@ class CouponController extends Controller
 
     public function store(CouponRequest $request, ValidateCouponPromocode $action)
     {
+        $this->authorize('Create-Coupon');
         $validatedData = $request->validated();
         $action->validatePromocode($request->promocode, $this->store);
         $coupon = $this->store->coupons()->create($validatedData);
@@ -40,7 +44,7 @@ class CouponController extends Controller
 
     public function show($couponId)
     {
-
+        $this->authorize('View-Coupon');
         $coupon = $this->storeService->findStoreModel($this->store, Captain::class, $couponId);
 
         return new CouponResource($coupon);
@@ -48,7 +52,7 @@ class CouponController extends Controller
 
     public function update(UpdateCouponRequest $request, $couponId, ValidateCouponPromocode $action)
     {
-
+        $this->authorize('Edit-Coupon');
         $coupon = $this->storeService->findStoreModel($this->store, Captain::class, $couponId);
         $action->validateExestingPromocode($couponId, $request->promocode, $this->store);
         $coupon->update($request->validated());
@@ -58,7 +62,7 @@ class CouponController extends Controller
 
     public function destroy($couponId)
     {
-
+        $this->authorize('Delete-Coupon');
         $coupon = $this->storeService->findStoreModel($this->store, Captain::class, $couponId);
         $coupon->delete();
